backend/requirements.txt
# Django==5.0.4
# djangorestframework==3.15.1
# djangorestframework-simplejwt==5.3.1
# django-cors-headers==4.3.1
# Pillow==10.3.0
# python-decouple==3.8
# psycopg2-binary==2.9.9
# python-dateutil==2.9.0

Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
python-decouple
psycopg2-binary
python-dateutil
